+++ package.json
@@ -10,11 +10,15 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.2",
     "axios": "^0.27.2",
     "glob": "^7.2.0",
     "react": "17.0.2",
     "react-native": "0.68.2",
-    "react-native-responsive-screen": "^1.4.2"
+    "react-native-responsive-screen": "^1.4.2",
+    "react-native-safe-area-context": "^4.2.5",
+    "react-native-screens": "^3.13.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
